device: yunzii-al66 / xinmeng-m66 (#86)
I have a yunzii-al66 board, and it is (I think) internally the same as
the yunzii-al71, but with 5 fewer keys. The `vendor_id` and `product_id`
are the same as the yunzi-al71. It follows that it's probably also the
same as the xinmeng-m66 (but i dont own this board). Writing also seemed
to work.