feature: device selection by report id (#103)
Enables correct device selection based on report-id found in the report
descriptor.

This is a feature I wanted to have in this tool from day one, but due to
both hidapi's lack of get_report_descriptor and rust lacking a robust
hid descriptor parsing library at the time, I haven't gotten around to
doing it... until now!

This will greatly simplify selecting the correct HID device on Windows
and hopefully make it less error-prone for most users, though this still
relies on knowing the correct interface number ahead of time.